+++ HISTORY.rst
@@ -17,3 +17,10 @@ History
 ------------------
 
 * bump up versions
+
+
+0.2.0 (2023-02-08)
+------------------
+
+* add eva (Extreme value analysis) module
+* fix bug in obtaining distribution parameters using optimization method